spill out fuse_column and fuse_type
