+++ src/query/service/tests/it/storages/fuse/operations/internal_column.rs
@@ -23,9 +23,9 @@ use common_expression::DataBlock;
 use common_sql::binder::INTERNAL_COLUMN_FACTORY;
 use common_sql::Planner;
 use common_storages_fuse::io::MetaReaders;
+use common_storages_fuse::FusePartInfo;
 use common_storages_fuse::FuseTable;
 use databend_query::interpreters::InterpreterFactory;
-use databend_query::storages::fuse::fuse_part::FusePartInfo;
 use databend_query::test_kits::table_test_fixture::execute_query;
 use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures::TryStreamExt;
