+++ src/query/storages/fuse/src/lib.rs
@@ -23,25 +23,30 @@
 #![recursion_limit = "256"]
 
 mod constants;
+mod fuse_column;
 mod fuse_lazy_part;
-pub mod fuse_part;
+mod fuse_part;
 mod fuse_table;
+mod fuse_type;
+mod metrics;
+
 pub mod io;
 pub mod operations;
 pub mod pruning;
 pub mod statistics;
 pub mod table_functions;
 
-mod metrics;
 use common_catalog::table::NavigationPoint;
 use common_catalog::table::Table;
 use common_catalog::table::TableStatistics;
 pub use common_catalog::table_context::TableContext;
 pub use constants::*;
+pub use fuse_column::FuseTableColumnStatisticsProvider;
 pub use fuse_lazy_part::FuseLazyPartInfo;
 pub use fuse_part::FusePartInfo;
-pub use fuse_table::FuseStorageFormat;
 pub use fuse_table::FuseTable;
+pub use fuse_type::FuseStorageFormat;
+pub use fuse_type::FuseTableType;
 pub use io::MergeIOReadResult;
 pub use pruning::SegmentLocation;
